C. Wood [#4]: Split 'Particle' struct into a separate module.
